requirements.txt
Flask==3.0.3
pymongo==4.4.0
python-dotenv==1.0.0
werkzeug==3.0.3
Pillow==10.4.0
reportlab==4.0.0
openai==1.35.7
